Refactoring

+++ football_scores/app/src/main/java/barqsoft/footballscores/ViewHolder.java
@@ -15,7 +15,7 @@ public class ViewHolder
     public TextView date;
     public ImageView home_crest;
     public ImageView away_crest;
-    public double match_id;
+    public int match_id;
     public ViewHolder(View view)
     {
         home_name = (TextView) view.findViewById(R.id.home_name);
